HV-2001 Add module-info generation via moditect
