+++ cdi/src/main/java/org/hibernate/validator/cdi/ValidationExtension.java
@@ -51,8 +51,8 @@ import org.hibernate.validator.cdi.internal.InheritedMethodsHelper;
 import org.hibernate.validator.cdi.internal.ValidationProviderHelper;
 import org.hibernate.validator.cdi.internal.ValidatorBean;
 import org.hibernate.validator.cdi.internal.ValidatorFactoryBean;
-import org.hibernate.validator.cdi.internal.interceptor.ValidationEnabledAnnotatedType;
-import org.hibernate.validator.cdi.internal.interceptor.ValidationInterceptor;
+import org.hibernate.validator.cdi.interceptor.internal.ValidationEnabledAnnotatedType;
+import org.hibernate.validator.cdi.interceptor.spi.ValidationInterceptor;
 import org.hibernate.validator.cdi.internal.util.GetterPropertySelectionStrategyHelper;
 import org.hibernate.validator.internal.util.Contracts;
 import org.hibernate.validator.internal.util.ExecutableHelper;
