+++ cdi/src/main/java/org/hibernate/validator/cdi/interceptor/internal/package-info.java
@@ -1,0 +1,11 @@
+/*
+ * Hibernate Validator, declare and validate application constraints
+ *
+ * License: Apache License, Version 2.0
+ * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
+ */
+
+/**
+ * <p>Classes for enabling Bean Validation method validation via a CDI interceptor.
+ */
+package org.hibernate.validator.cdi.interceptor.internal;
