--- cdi/src/main/java/org/hibernate/validator/cdi/internal/interceptor/package-info.java
@@ -1,11 +1,0 @@
-/*
- * Hibernate Validator, declare and validate application constraints
- *
- * License: Apache License, Version 2.0
- * See the license.txt file in the root directory or <http://www.apache.org/licenses/LICENSE-2.0>.
- */
-
-/**
- * <p>Classes for enabling Bean Validation method validation via a CDI interceptor.
- */
-package org.hibernate.validator.cdi.internal.interceptor;
