+++ cdi/src/test/java/org/hibernate/validator/test/cdi/internal/InjectingConstraintValidatorFactoryTest.java
@@ -21,7 +21,7 @@ import jakarta.validation.ConstraintValidator;
 import jakarta.validation.ConstraintValidatorContext;
 import jakarta.validation.constraints.Min;
 
-import org.hibernate.validator.cdi.internal.InjectingConstraintValidatorFactory;
+import org.hibernate.validator.cdi.spi.InjectingConstraintValidatorFactory;
 import org.testng.annotations.BeforeClass;
 import org.testng.annotations.Test;
