+++ java/cdi/src/main/resources/META-INF/services/jakarta.validation.ConstraintValidator
@@ -1,0 +1,1 @@
+org.hibernate.validator.integrationtest.java.module.cdi.constraint.CarServiceConstraint$Validator
